fix: add user checking for booking to sure that booking belongs to a particular user

+++ app/services/bookings.py
@@ -1,7 +1,9 @@
 from decimal import Decimal
 
+from sqlalchemy.sql.functions import current_user
+
 from app.exceptions.base import FlightNotAvailableForBookingException, SeatsNotAvailableException, \
-    BookingNotFoundException, ObjectNotFoundException, PassengerNotFoundException
+    BookingNotFoundException, ObjectNotFoundException, PassengerNotFoundException, ForbiddenBookingException
 from app.helpers.booking_status import BookingStatus
 from app.helpers.flight_status import FlightStatus
 from app.helpers.seat_status import SeatStatus
@@ -83,8 +85,10 @@ class BookingService(BaseService):
         await self.db.commit()
         return {"detail": "Бронирование успешно удалено"}
 
-    async def delete_passenger_in_booking(self, passenger_id: int, booking_id: int):
+    async def delete_passenger_in_booking(self, passenger_id: int, booking_id: int, user_id: int):
         booking = await self.get_booking_or_404(booking_id)
+        if booking.user_id != user_id:
+            raise ForbiddenBookingException
         if len(booking.passengers) == 1:
             return await self.delete_booking_fully(booking_id)
         passenger = next((p for p in booking.passengers if p.id == passenger_id), None)
